Move encoders into separate functions in generated module
Objects can have bidirectional relationships.
Therefore their relationship cannot be expressed
using a acyclic graph and therefore we cannot
create the encoders inline because it causes
infinite recursion. I.e. if User has
a relationship to User the depth of users can
be infinitely big.

That's why separate encoders are introduced.
They are defined according to the following scheme:
json_of_TYPE where TYPE is type name from the schema

This commit screws up error locations though,
but it will be fixed in the following commits.