Use interface files to assert the generated type of a query
This also switches to Reason syntax for tests since refmt makes it
a bit easier to write the type signatures.

It also removes the execute step - at least for now - of the tests.
This effectively means that we only validate the signature, not that
the parser actually works. Those tests are yet to be done...