Add support for parsing selection sets into records
Adding the @bsRecord directive to any field with a selection
set now causes the equivalent code of {field: value} be generated.

This of course requires an existing record type with the matching types
to work, but it will generate type errors if anything mismatches.

By using records instead of Js.t objects you get a slightly more
efficient in-memory representation, but the two biggest upsides are
that you can pattern match on the result and create updated records
with the spread/`with` syntax.