Properly handle @skip/@include directives
If you put @skip or @include on any field, it will be forced into
an optional, even if the underlying field is non-nullable.

Putting @skip/@include on a nullable field will however only produce
one level of optionals, so there's no way of determining whether a
field was null or omitted.